+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.0.4
+* Fix issue when milestones were rendered twice
+* Show customized legend name in the tooltips
+* Packages update
+
 ## 3.0.3
 * Render "Collapse all" / "Expand all" even if "Category labels" is disabled